Restructure features and common folders
