+++ src/app/App.js
@@ -1,5 +1,5 @@
 import { BrowserRouter, Switch, Route } from 'react-router-dom';
-import Header from '../common/Header';
+import Header from '../features/header/Header';
 import Books from '../features/books/Books';
 import Categories from '../features/categories/Categories';
 import './scss/styles.scss';
