+++ src/features/header/Header.js
@@ -1,6 +1,6 @@
 import PersonIcon from '@material-ui/icons/Person';
-import WebsiteTitle from './WebsiteTitle';
-import Navigation from './Navigation';
+import WebsiteTitle from '../../common/components/WebsiteTitle';
+import Navigation from '../../common/components/Navigation';
 import styles from './Header.module.scss';
 
 const Header = () => (
