+++ src/features/header/Header.module.scss
@@ -1,4 +1,4 @@
-@import '../app/scss/flex';
+@import '../../app/scss/flex';
 
 .header {
   @extend %flex-row;
